Move resource search functionality to default repository
